Add MF_PLUGIN_PATH to test environment
A couple of tkeyvfs tests fail due to use of local bash scripts
and subsequent failure to pass down DYLD_LIBRARY_PATH on macOS.
An execption is thrown by *messagefacility* because the script
wraps a small executable indirectly linked to MF_MessageLogger.
The internals of MF_MessageLogger require, since 2.2.0, the presence
of the dynamic loader path in the environment, even if it is not
directly used.

In the ALT_CMAKE test environment, set MF_PLUGIN_PATH to the paths
to art libraries and MF_MessageLogger. This ensures consistent
behaviour across tests without having toe modify scripts directly.